refactor: fedimint-gateway-common
