+++ devimint/Cargo.toml
@@ -25,7 +25,7 @@ fedimint-bitcoind = { workspace = true }
 fedimint-client = { workspace = true }
 fedimint-client-module = { workspace = true }
 fedimint-core = { workspace = true }
-fedimint-gateway-server = { workspace = true }
+fedimint-gateway-common = { workspace = true }
 fedimint-lightning = { workspace = true }
 fedimint-ln-client = { workspace = true, features = ["cli"] }
 fedimint-ln-server = { workspace = true }
