+++ fedimint-testing/Cargo.toml
@@ -12,7 +12,7 @@ repository = { workspace = true }
 rustc-args = ["--cfg", "tokio_unstable"]
 
 [features]
-tor = ["fedimint-gateway-server/tor"]
+tor = ["fedimint-gateway-server/tor", "fedimint-gateway-common/tor"]
 
 [lib]
 name = "fedimint_testing"
@@ -30,6 +30,7 @@ fedimint-bitcoind = { workspace = true }
 fedimint-client = { workspace = true }
 fedimint-client-module = { workspace = true }
 fedimint-core = { workspace = true }
+fedimint-gateway-common = { workspace = true }
 fedimint-gateway-server = { package = "fedimint-gateway-server", path = "../gateway/fedimint-gateway-server", version = "=0.7.0-alpha" }
 fedimint-lightning = { package = "fedimint-lightning", path = "../gateway/fedimint-lightning", version = "=0.7.0-alpha" }
 fedimint-ln-common = { workspace = true }
