+++ gateway/cli/src/ecash_commands.rs
@@ -3,11 +3,11 @@ use bitcoin::Address;
 use clap::Subcommand;
 use fedimint_core::config::FederationId;
 use fedimint_core::{Amount, BitcoinAmountOrAll};
-use fedimint_gateway_server::rpc::rpc_client::GatewayRpcClient;
-use fedimint_gateway_server::rpc::{
+use fedimint_gateway_common::{
     BackupPayload, DepositAddressPayload, DepositAddressRecheckPayload, ReceiveEcashPayload,
     SpendEcashPayload, WithdrawPayload,
 };
+use fedimint_gateway_server::rpc::rpc_client::GatewayRpcClient;
 use fedimint_mint_client::OOBNotes;
 
 use crate::print_response;
