+++ gateway/cli/src/main.rs
@@ -10,8 +10,8 @@ use clap::{CommandFactory, Parser, Subcommand};
 use config_commands::ConfigCommands;
 use ecash_commands::EcashCommands;
 use fedimint_core::util::SafeUrl;
+use fedimint_gateway_common::V1_API_ENDPOINT;
 use fedimint_gateway_server::rpc::rpc_client::GatewayRpcClient;
-use fedimint_gateway_server::rpc::V1_API_ENDPOINT;
 use fedimint_logging::TracingSetup;
 use general_commands::GeneralCommands;
 use lightning_commands::LightningCommands;
