+++ gateway/fedimint-gateway-server/src/client.rs
@@ -13,10 +13,11 @@ use fedimint_core::core::ModuleKind;
 use fedimint_core::db::{Database, IDatabaseTransactionOpsCoreTyped};
 use fedimint_core::module::registry::ModuleDecoderRegistry;
 use fedimint_derive_secret::{ChildId, DerivableSecret};
+use fedimint_gateway_common::FederationConfig;
 use fedimint_gw_client::GatewayClientInit;
 use fedimint_gwv2_client::GatewayClientInitV2;
 
-use crate::db::{FederationConfig, GatewayDbExt};
+use crate::db::GatewayDbExt;
 use crate::error::AdminGatewayError;
 use crate::{AdminResult, Gateway};
