+++ gateway/integration_tests/Cargo.toml
@@ -18,7 +18,7 @@ anyhow = { workspace = true }
 clap = { workspace = true }
 devimint = { workspace = true }
 fedimint-core = { workspace = true }
-fedimint-gateway-server = { workspace = true }
+fedimint-gateway-common = { workspace = true }
 fedimint-testing = { workspace = true }
 itertools = { workspace = true }
 serde_json = { workspace = true }
